requirements.txt
Flask==2.2.2
requests==2.31.0
python-dotenv==0.19.2
gunicorn==20.0.4
google-auth==2.0.2

